half regex init
